Added close button to all modals
